# openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: backend@callverse.local
    name: CallVerse backend
  description: |-
    Backend for CallVerse, a digital twin of a retail bank's customer relation center.

    **Two modes, one API.** Every endpoint behaves identically in live mode (a real person on the customer portal) and simulation mode (synthetic customers driving a reinforcement-learning experiment). The only difference is the source of the customers, expressed in the data as `customer.is_simulated` and `conversation.run_id`.

    **The backend is the authority on business rules.** A commercial credit above an advisor's ceiling is refused here, whoever asks.

    **Conventions.** Routes are versioned under `/api/v1`. UUIDs are exposed publicly and sequence identifiers never are. All timestamps are ISO-8601 UTC. Every failure returns the same envelope — `timestamp`, `status`, `code`, `message`, `path` — and clients branch on `code`, never on `message`.
  license:
    name: Academic project - 3iL / ESPRIT
  title: CallVerse Backend API
  version: 0.1.0-SNAPSHOT
servers:
- description: Local development
  url: http://localhost:8080
- description: Relative to the deployed host
  url: /
...
paths:
...
  /api/v1/cards/{id}/block:
    post:
      description: "Blocks an ACTIVE card with a reason and the server's time. Idempotent:\
        \ an already BLOCKED card is returned unchanged, keeping its first reason\
        \ and time. An EXPIRED or CANCELLED card is 409 INVALID_STATE_TRANSITION;\
        \ an unknown card 404. Advisors and supervisors only."
      operationId: blockCard
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockCardRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "Rejected before any business rule ran: the body is not valid\
            \ JSON (code MALFORMED_REQUEST), a parameter or field broke a rule (code\
            \ VALIDATION_FAILED), or a use-case precondition was unmet."
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "Not authenticated. code UNAUTHENTICATED: no bearer token,\
            \ or one that is expired, malformed or wrongly signed - all indistinguishable,\
            \ and the signal to refresh or log in again. code INVALID_CREDENTIALS:\
            \ login refused; an unknown email and a wrong password are deliberately\
            \ indistinguishable."
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "Authenticated, but not permitted: code ACCESS_DENIED. Re-authenticating\
            \ will not help, so clients must not refresh on it."
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "No such route, or a referenced resource does not exist."
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: A business rule refused the operation. The request was well-formed;
            the current state of the business forbids it.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unexpected failure. message is deliberately generic and carries
            no internal detail; see the server log for the stack trace.
      security:
      - bearerAuth: []
      summary: Block a payment card
      tags:
      - Cards
  /api/v1/conversations/{id}/escalations:
    post:
      description: "Raised by ADVISOR. Idempotent: a conversation that already has\
        \ a pending escalation returns it with 200 instead of creating a second. Only\
        \ an ACTIVE conversation can be escalated, else 409 INVALID_STATE_TRANSITION;\
        \ an unknown conversation is 404. Advisors only: a supervisor is who an escalation\
        \ goes to."
      operationId: escalateConversation
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscalationRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscalationResponse"
          description: A pending escalation already existed and is returned
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscalationResponse"
          description: Escalation created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "Rejected before any business rule ran: the body is not valid\
            \ JSON (code MALFORMED_REQUEST), a parameter or field broke a rule (code\
            \ VALIDATION_FAILED), or a use-case precondition was unmet."
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "Not authenticated. code UNAUTHENTICATED: no bearer token,\
            \ or one that is expired, malformed or wrongly signed - all indistinguishable,\
            \ and the signal to refresh or log in again. code INVALID_CREDENTIALS:\
            \ login refused; an unknown email and a wrong password are deliberately\
            \ indistinguishable."
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "Authenticated, but not permitted: code ACCESS_DENIED. Re-authenticating\
            \ will not help, so clients must not refresh on it."
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "No such route, or a referenced resource does not exist."
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: A business rule refused the operation. The request was well-formed;
            the current state of the business forbids it.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unexpected failure. message is deliberately generic and carries
            no internal detail; see the server log for the stack trace.
      security:
      - bearerAuth: []
      summary: Escalate a conversation to a supervisor
      tags:
      - Conversations
...
  /api/v1/kb/articles:
    get:
      description: "Case-insensitive search over published articles; q is 2 to 100\
        \ characters, limit 1 to 10 (default 5). Staff only."
      operationId: searchKnowledgeArticles
      parameters:
      - in: query
        name: q
        required: true
        schema:
          type: string
      - in: query
        name: limit
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeArticlesResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "Rejected before any business rule ran: the body is not valid\
            \ JSON (code MALFORMED_REQUEST), a parameter or field broke a rule (code\
            \ VALIDATION_FAILED), or a use-case precondition was unmet."
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "Not authenticated. code UNAUTHENTICATED: no bearer token,\
            \ or one that is expired, malformed or wrongly signed - all indistinguishable,\
            \ and the signal to refresh or log in again. code INVALID_CREDENTIALS:\
            \ login refused; an unknown email and a wrong password are deliberately\
            \ indistinguishable."
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "Authenticated, but not permitted: code ACCESS_DENIED. Re-authenticating\
            \ will not help, so clients must not refresh on it."
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "No such route, or a referenced resource does not exist."
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: A business rule refused the operation. The request was well-formed;
            the current state of the business forbids it.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unexpected failure. message is deliberately generic and carries
            no internal detail; see the server log for the stack trace.
      security:
      - bearerAuth: []
      summary: Search published articles
      tags:
      - Knowledge base
...
  /api/v1/tickets:
    post:
      description: "Always created OPEN; severity 1 to 5, default 3. When conversationId\
        \ is given it must belong to customerId, else 400 CONVERSATION_CUSTOMER_MISMATCH.\
        \ Unknown customer or conversation: 404. Advisors and supervisors only."
      operationId: openTicket
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenTicketRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketResponse"
          description: Ticket created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "Rejected before any business rule ran: the body is not valid\
            \ JSON (code MALFORMED_REQUEST), a parameter or field broke a rule (code\
            \ VALIDATION_FAILED), or a use-case precondition was unmet."
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "Not authenticated. code UNAUTHENTICATED: no bearer token,\
            \ or one that is expired, malformed or wrongly signed - all indistinguishable,\
            \ and the signal to refresh or log in again. code INVALID_CREDENTIALS:\
            \ login refused; an unknown email and a wrong password are deliberately\
            \ indistinguishable."
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "Authenticated, but not permitted: code ACCESS_DENIED. Re-authenticating\
            \ will not help, so clients must not refresh on it."
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "No such route, or a referenced resource does not exist."
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: A business rule refused the operation. The request was well-formed;
            the current state of the business forbids it.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unexpected failure. message is deliberately generic and carries
            no internal detail; see the server log for the stack trace.
      security:
      - bearerAuth: []
      summary: Open a support ticket
      tags:
      - Tickets
components:
  schemas:
    BlockCardRequest:
      type: object
      properties:
        reason:
          type: string
          enum:
          - LOST
          - STOLEN
          - FRAUD_SUSPECTED
          - CUSTOMER_REQUEST
          example: STOLEN
      required:
      - reason
    CardResponse:
      type: object
      description: A payment card
      properties:
        accountId:
          type: string
          format: uuid
        blockReason:
          type: string
          enum:
          - LOST
          - STOLEN
          - FRAUD_SUSPECTED
          - CUSTOMER_REQUEST
          example: STOLEN
          nullable: true
        blockedAt:
          type: string
          format: date-time
          description: Set when the card was blocked
          nullable: true
        expiresOn:
          type: string
          format: date
        id:
          type: string
          format: uuid
        network:
          type: string
          enum:
          - VISA
          - MASTERCARD
          example: VISA
        panLast4:
          type: string
          example: "4242"
        status:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
          - CANCELLED
          example: BLOCKED
        type:
          type: string
          enum:
          - DEBIT
          - CREDIT
          example: DEBIT
      required:
      - accountId
      - expiresOn
      - id
      - network
      - panLast4
      - status
      - type
...
    ErrorResponse:
      type: object
      description: Standard error envelope returned by every CallVerse endpoint
      properties:
        code:
          type: string
          example: RESOURCE_NOT_FOUND
        message:
          type: string
          example: Customer 'a3f1...' was not found
        path:
          type: string
          example: /api/v1/customers/a3f1
        status:
          type: integer
          format: int32
          example: 404
        timestamp:
          type: string
          format: date-time
          example: 2026-09-14T08:31:07.412Z
      required:
      - code
      - message
      - path
      - status
      - timestamp
    EscalationRequest:
      type: object
      properties:
        reason:
          type: string
          example: "Fraude suspectee, besoin d'un superviseur"
          maxLength: 255
          minLength: 0
      required:
      - reason
    EscalationResponse:
      type: object
      description: An escalation of a conversation to a supervisor
      properties:
        conversationId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        raisedBy:
          type: string
          enum:
          - ADVISOR
          - AI
          - RULE
          example: ADVISOR
        reason:
          type: string
        status:
          type: string
          example: PENDING
      required:
      - conversationId
      - createdAt
      - id
      - raisedBy
      - reason
      - status
...
    KnowledgeArticle:
      type: object
      description: One published article
      properties:
        category:
          type: string
          example: CARDS
        content:
          type: string
        id:
          type: string
          format: uuid
        tags:
          type: array
          items:
            type: string
        title:
          type: string
          example: Faire opposition a votre carte
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int32
          example: 1
      required:
      - category
      - content
      - id
      - tags
      - title
      - updatedAt
      - version
    KnowledgeArticlesResponse:
      type: object
      description: Published articles matching a search
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeArticle"
        query:
          type: string
          example: opposition carte
      required:
      - articles
      - query
...
    OpenTicketRequest:
      type: object
      properties:
        category:
          type: string
          example: FRAUD
          maxLength: 30
          minLength: 0
          pattern: "[A-Z][A-Z_]*"
        conversationId:
          type: string
          format: uuid
          description: "When given, must belong to customerId"
          nullable: true
        customerId:
          type: string
          format: uuid
        description:
          type: string
          maxLength: 4000
          minLength: 0
          nullable: true
        severity:
          type: integer
          format: int32
          description: "1 to 5, default 3"
          example: 3
          maximum: 5
          minimum: 1
          nullable: true
        title:
          type: string
          example: Paiements inconnus a l'etranger
          maxLength: 200
          minLength: 0
      required:
      - category
      - customerId
      - title
...
    TicketResponse:
      type: object
      description: A support ticket
      properties:
        category:
          type: string
          example: FRAUD
        conversationId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        customerId:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        severity:
          type: integer
          format: int32
          description: 1 (most severe) to 5
          example: 2
        status:
          type: string
          example: OPEN
        title:
          type: string
          example: Paiements inconnus a l'etranger
      required:
      - category
      - createdAt
      - customerId
      - id
      - severity
      - status
      - title
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: |-
        User access token from `POST /api/v1/auth/login`. Carries the account identifier and one role of CUSTOMER, ADVISOR, SUPERVISOR or ADMIN.

        HS512, valid for one hour by default. Missing, expired or invalid: 401 UNAUTHENTICATED.
      scheme: bearer
      type: http